requirements.txt
rich
argparse
lxml
selenium
pillow